--- src/main/java/gift/Exception/InvalidTokenException.java
@@ -1,7 +1,0 @@
-package gift.Exception;
-
-public class InvalidTokenException extends BusinessException {
-    public InvalidTokenException(String message) {
-        super(message);
-    }
-}
