--- src/main/java/gift/Exception/NotValidTokenException.java
@@ -1,7 +1,0 @@
-package gift.Exception;
-
-public class NotValidTokenException extends BusinessException {
-    public NotValidTokenException(String message) {
-        super(message);
-    }
-}
