--- src/main/java/gift/Exception/NullTokenException.java
@@ -1,7 +1,0 @@
-package gift.Exception;
-
-public class NullTokenException extends BusinessException {
-    public NullTokenException(String message) {
-        super(message);
-    }
-}
