+++ src/main/java/gift/Exception/TokenException/InvalidTokenException.java
@@ -1,0 +1,7 @@
+package gift.Exception.TokenException;
+
+public class InvalidTokenException extends TokenException {
+    public InvalidTokenException(String message) {
+        super(message);
+    }
+}
