+++ src/main/java/gift/Exception/TokenException/NotValidTokenException.java
@@ -1,0 +1,7 @@
+package gift.Exception.TokenException;
+
+public class NotValidTokenException extends TokenException {
+    public NotValidTokenException(String message) {
+        super(message);
+    }
+}
