+++ src/main/java/gift/Exception/TokenException/NullTokenException.java
@@ -1,0 +1,7 @@
+package gift.Exception.TokenException;
+
+public class NullTokenException extends TokenException {
+    public NullTokenException(String message) {
+        super(message);
+    }
+}
