+++ src/main/java/gift/Exception/TokenException/TokenException.java
@@ -1,0 +1,8 @@
+package gift.Exception.TokenException;
+
+public class TokenException  extends RuntimeException{
+    public TokenException(String message) {
+        super(message);
+    }
+}
+
