+++ src/main/java/gift/Exception/TokenException/TokenExpiredException.java
@@ -1,0 +1,7 @@
+package gift.Exception.TokenException;
+
+public class TokenExpiredException extends TokenException {
+    public TokenExpiredException(String message) {
+        super(message);
+    }
+}
