--- src/main/java/gift/Exception/TokenExpiredException.java
@@ -1,7 +1,0 @@
-package gift.Exception;
-
-public class TokenExpiredException extends BusinessException {
-    public TokenExpiredException(String message) {
-        super(message);
-    }
-}
